refactor(server): isolate hosted to local authority transfer
